backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic==2.5.1
geopandas==0.13.2
fiona==1.9.4
shapely==2.0.2
pyproj==3.6.1
openpyxl==3.1.2
earthengine-api==0.1.375
celery==5.3.5
redis==5.0.1
boto3==1.29.3
python-dotenv==1.0.0
mangum==0.17.0